requirements.txt
mesa
psycopg2-binary
python-dotenv
pandas
catboost
joblib
numpy
scikit-learn
PySide6
pyqtgraph
fastapi
uvicorn
